feat(prime): add A91B2 charging station + shared base + post-connect refactor
- Extract PrimeUsbCharger (six-USB-port telemetry + serial + the ca00 command),
  shared by both Prime chargers; PrimeCharger250w becomes a thin subclass.
- Add PrimeChargingStation240w (A91B2): USB ports via the base plus the two
  AC-outlet switch states (ac_1_switch/ac_2_switch); decode validated against the
  anker-solix-api A91B2 map. Live telemetry pending its post-ECDH enable step.
- Refactor negotiation: the device's trailing 4822/4827 after the stage-5 ECDH
  are acks, not stages needing a response. Move the registration (4027) and
  subscribe (4200) to _post_connect, mirroring the gen-2 power stations' 4100.
  Negotiation writes use write-with-response for reliability on weaker links.
- PrimeChargingStation240w overrides _post_connect to a no-op: it rejects the
  charger's 4027 registration (dropping the link) and its own enable sequence is
  unknown (needs an Anker-app handshake capture), so it negotiates and stays
  connected but does not stream yet.

Confirmed on an A2345: negotiates, registers + subscribes from _post_connect, and
streams ca00 telemetry, decoding per-port values matching the cloud.
