+++ SolixBLE/__init__.py
@@ -17,7 +17,9 @@ from .devices import (
     MagGo3in1,
     PrimeCharger160w,
     PrimeCharger250w,
+    PrimeChargingStation240w,
     PrimePowerBank20k,
+    PrimeUsbCharger,
     Solarbank2,
     Solarbank3,
 )
@@ -47,7 +49,9 @@ __all__ = [
     "Solarbank3",
     "PrimeCharger160w",
     "PrimeCharger250w",
+    "PrimeChargingStation240w",
     "PrimePowerBank20k",
+    "PrimeUsbCharger",
     "MagGo3in1",
     "Generic",
     "ChargingStatus",
